+++ click_extra/tests/test_logging.py
@@ -15,8 +15,6 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
-import re
-
 import click
 import pytest
 
@@ -86,13 +84,11 @@ def test_integrated_verbosity_option(invoke, level):
     assert result.exit_code == 0
     assert result.output == "It works!\nRun command #1...\n"
     if level == "DEBUG":
-        assert re.fullmatch(
-            r"\x1b\[34mdebug: \x1b\[0mVerbosity set to DEBUG.\n"
-            r"\x1b\[34mdebug: \x1b\[0mLoad configuration from \S+config.toml\n"
-            r"\x1b\[34mdebug: \x1b\[0mConfiguration not found at \S+config.toml\n"
-            r"\x1b\[34mdebug: \x1b\[0mIgnore configuration file.\n"
-            r"\x1b\[34mdebug: \x1b\[0mLoaded configuration: {}\n",
-            result.stderr,
+        assert result.stderr == (
+            "\x1b[34mdebug: \x1b[0mVerbosity set to DEBUG.\n"
+            "\x1b[34mdebug: \x1b[0mSearch for configuration in default location...\n"
+            "\x1b[34mdebug: \x1b[0mNo default configuration found.\n"
+            "\x1b[34mdebug: \x1b[0mNo configuration provided.\n"
         )
     else:
         assert not result.stderr