+++ changelog.md
@@ -6,6 +6,10 @@
 This version is not released yet and is under active development.
 ```
 
+- Search all supported formats in default location if configuration file not
+  provided.
+- Print configuration file default location in help screens.
+
 ## [1.4.1 (2022-02-13)](https://github.com/kdeldycke/click-extra/compare/v1.4.0...v1.4.1)
 
 - Add new external workflow to modernize Python code.